upgrade webpack, add ts support
migrate from webpack v4 to v5. this significantly changes the current webpack config. removed cross-env since it was no longer needed. now we use DEV property to read mode and set NODE_ENV accordingly. configuration now uses terset plugin instead of uglify since it is supposed to be faster and more efficient. it is also part of webpack as a default minimizer so i guess it's time to start using that. removed splitchunks configuration because the defaults seem good to be in a boilerplate. removed statsconfig because i don't remember why i thought i needed it in every project. added clean-webpack-plugin that is supposed to clean build folders. that's neat. small changes here and there based on the v4 vs v5 differences. for example, don't need url loader now, just use asset type. one major issue is webpack seems to be unable to resolve src directory unless an index file is present. this needs to be fixed.

Signed-off-by: rishichawda <rishichawda@users.noreply.github.com>